refactor: extract Factory.Test from Servcie.Test
